Feat: 서치파라미터로 모달기능 구현 테스트 완료
