+++ app/globals.css
@@ -7,6 +7,7 @@
 }
 
 body {
+  position: relative;
   margin: 0;
 }
