+++ app/modal/@modal/(.)newDashboard/default.tsx
@@ -1,0 +1,5 @@
+function Default() {
+  return null;
+}
+
+export default Default;
